Split sled_storage codes - add error and store_mut modules,
In sled_storage/,
Detach error codes from mod.rs and move to error.rs.
Detach impl StoreMut from store.rs and move to store_mut.rs.
