+++ src/storages/sled_storage/mod.rs
@@ -1,59 +1,15 @@
+mod alter_table;
+mod error;
+mod store;
+mod store_mut;
+#[cfg(not(feature = "alter-table"))]
+impl crate::AlterTable for SledStorage {}
+
 use sled::{self, Config, Db};
 use std::convert::TryFrom;
-use std::str;
-use thiserror::Error as ThisError;
 
-#[cfg(not(feature = "alter-table"))]
-use crate::AlterTable;
-#[cfg(feature = "alter-table")]
-use crate::AlterTableError;
 use crate::{Error, Result, Schema};
-
-#[cfg(feature = "alter-table")]
-mod alter_table;
-#[cfg(not(feature = "alter-table"))]
-impl AlterTable for SledStorage {}
-
-mod store;
-
-#[derive(ThisError, Debug)]
-enum StorageError {
-    #[cfg(feature = "alter-table")]
-    #[error(transparent)]
-    AlterTable(#[from] AlterTableError),
-
-    #[error(transparent)]
-    Sled(#[from] sled::Error),
-    #[error(transparent)]
-    Bincode(#[from] bincode::Error),
-    #[error(transparent)]
-    Str(#[from] str::Utf8Error),
-}
-
-impl Into<Error> for StorageError {
-    fn into(self) -> Error {
-        use StorageError::*;
-
-        match self {
-            Sled(e) => Error::Storage(Box::new(e)),
-            Bincode(e) => Error::Storage(e),
-            Str(e) => Error::Storage(Box::new(e)),
-
-            #[cfg(feature = "alter-table")]
-            AlterTable(e) => e.into(),
-        }
-    }
-}
-
-fn err_into<E>(e: E) -> Error
-where
-    E: Into<StorageError>,
-{
-    let e: StorageError = e.into();
-    let e: Error = e.into();
-
-    e
-}
+use error::err_into;
 
 #[derive(Debug, Clone)]
 pub struct SledStorage {
